\author{Hyuna Yang and Keith Sheppard}

\name{plotMouseDivArrayImage}
\alias{plotMouseDivArrayImage}

\title{.CEL file image plot}

\description{
  This function generates a image plot of a mouse diversity .CEL file.
}

\usage{
plotMouseDivArrayImage(celFilename)
}

\arguments{
  \item{celFilename}{Name of a .CEL file that you want to draw image plot of.}
}

\examples{
\dontrun{
plotMouseDivArrayImage("../../../celFiles/SNP_mDIV_A8-8_081308.CEL")
}}


\keyword{plot}
