\author{ Hyuna Yang }

\name{MouseDivGeno-internal}

\alias{ccstrans}
\alias{computegender}
\alias{test123}
\alias{test12}
\alias{vdist}
\alias{E.step3}
\alias{M.step3}
\alias{E.step2}
\alias{M.step2}

\usage{
ccstrans(a,b,k=4)
computegender(intX,intY, autoint
test123 (tscore2, tscore3, nm, tgeno2, tgeno3, rmid, thres, iig, nsize)
test12( tscore2, nm, tgeno, rmid, thres, iig, nsize)
vdist(ta,tb,t2)
E.step3(theta,data)
M.step3(T,data)
E.step2(theta,data)
M.step2(T,data)
}

\title{Internal MouseDivGeno functions}

\description{
  Internal MouseDivGeno functions. These are generally not to be called by
  the user. 
}


\keyword{internal}
