\author{Hyuna Yang}

\name{plotMouseDivArrayImage}
\alias{plotMouseDivArrayImage}

\title{.CEL file image plot}

\description{
  This function generates a image plot of a mouse diversity .CEL file.
}

\usage{
plotMouseDivArrayImage(celFilename)
}

\arguments{
  \item{celFilename}{Name of a .CEL file that you want to draw image plot of.}
}

\examples{
\dontrun{
plotMouseDivArrayImage("~/projects/cel/SNPex1.CEL")
}}


\keyword{plot}
